convert data drift into real-time events (week 9 day 7)
- Implemented full monitoring pipeline from drift detection to event triggering
- Introduced distribution shift scenarios to validate alert behavior
- Enabled Prometheus + Alertmanager integration with webhook routing
- Demonstrated drift escalation: stable → warning → critical
- Added end-to-end demo proof and logs

+++ src/drift/worker.py
@@ -7,7 +7,7 @@ from src.drift.detector import compute_drift, compute_confidence_shift, compute_
 from src.drift.storage.drift_metrics_repository import insert_drift_metric
 MIN_SAMPLES = 10
 
-def compute_time_windows(window_hours: int):
+def compute_time_windows(window_hours: float):
     now = datetime.now()
 
     cur_end = now
@@ -18,7 +18,7 @@ def compute_time_windows(window_hours: int):
 
     return ref_start, ref_end, cur_start, cur_end
 
-def run_drift_job(window_hours: int = 1) -> dict:
+def run_drift_job(window_hours: float = 1.0) -> dict:
     store = get_drift_store()
 
     # define windows
